feat(PPDSC-2415): feat-2415-added promise
